requirements.txt
cffi==1.15.0
click==8.0.1
cryptography==37.0.2
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
peewee==3.14.10
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.17.1
Werkzeug==2.0.1
flask_cors==4.0.0
jsonify==0.5